{"schema":"decodex/commit/1","summary":"Split agent evidence modules","authority":"manual"}
