+++ apps/decodex/src/orchestrator/agent_evidence/files.rs
@@ -1,9 +1,20 @@
-use super::{
+use std::{
+	collections::{self, BTreeMap},
+	fs::{self, OpenOptions},
+	io::Write as _,
+	path::{Path, PathBuf},
+	process,
+};
+
+use serde::Serialize;
+use time::OffsetDateTime;
+
+use crate::orchestrator::agent_evidence::{
 	AGENT_BLOCKER_SNAPSHOT_SCHEMA, AGENT_EVIDENCE_EVENT_SCHEMA, AgentBlocker, AgentBlockerSnapshot,
 	AgentEvidenceEvent, AgentEvidenceFileWriteContext, AgentEvidenceSource, AgentHandoffIndex,
-	AgentRunCapsule, AgentRunCapsuleRef, BTreeMap, OffsetDateTime, OpenOptions, Path, PathBuf,
-	Result, Serialize, Write, collections, eyre, fs, process,
+	AgentRunCapsule, AgentRunCapsuleRef,
 };
+use crate::prelude::{Result, eyre};
 
 pub(in crate::orchestrator) fn write_agent_evidence_files(
 	context: &AgentEvidenceFileWriteContext<'_>,
