+++ apps/decodex/src/orchestrator/agent_evidence/snapshot.rs
@@ -1,29 +1,32 @@
-use super::{
-	AGENT_HANDOFF_INDEX_SCHEMA, AgentEvidenceFileWriteContext, AgentEvidenceProjectView,
+use crate::orchestrator::agent_evidence::{
+	self, AGENT_HANDOFF_INDEX_SCHEMA, AgentEvidenceFileWriteContext, AgentEvidenceProjectView,
 	AgentEvidenceSource, AgentEvidenceSummary, AgentEvidenceWriteResult, AgentHandoffIndex,
 	BLOCKERS_DIR_NAME, EVENTS_FILE_NAME, HANDOFF_INDEX_FILE_NAME, OperatorStatusSnapshot,
-	RUNS_DIR_NAME, Result, agent_connector_backoff, agent_evidence_project_ids,
-	agent_recovery_contract, agent_recovery_worktree, build_agent_blockers, build_run_capsules,
-	current_month_bucket, current_timestamp, run_capsule_ref, runtime, write_agent_evidence_files,
+	RUNS_DIR_NAME, Result, agent_connector_backoff, agent_recovery_contract, run_capsule_ref,
+	runtime,
 };
 
 pub(in crate::orchestrator) fn write_agent_evidence_snapshot(
 	snapshot: &OperatorStatusSnapshot,
 	source: AgentEvidenceSource,
 ) -> Result<Vec<AgentEvidenceWriteResult>> {
-	let generated_at = current_timestamp();
-	let month_bucket = current_month_bucket();
+	let generated_at = agent_evidence::current_timestamp();
+	let month_bucket = agent_evidence::current_month_bucket();
 	let mut results = Vec::new();
 
-	for project_id in agent_evidence_project_ids(snapshot) {
+	for project_id in agent_evidence::agent_evidence_project_ids(snapshot) {
 		let service_root = runtime::agent_evidence_dir()?.join(&project_id);
 		let handoff_index_path = service_root.join(HANDOFF_INDEX_FILE_NAME);
 		let blockers_dir = service_root.join(BLOCKERS_DIR_NAME);
 		let runs_dir = service_root.join(RUNS_DIR_NAME);
 		let events_path = service_root.join(EVENTS_FILE_NAME);
 		let project_view = AgentEvidenceProjectView::from_snapshot(snapshot, &project_id);
-		let mut run_capsules =
-			build_run_capsules(&project_view, &generated_at, &runs_dir, &month_bucket);
+		let mut run_capsules = agent_evidence::build_run_capsules(
+			&project_view,
+			&generated_at,
+			&runs_dir,
+			&month_bucket,
+		);
 
 		run_capsules.sort_by(|left, right| {
 			left.issue_identifier
@@ -34,11 +37,12 @@ pub(in crate::orchestrator) fn write_agent_evidence_snapshot(
 		});
 
 		let run_refs = run_capsules.iter().map(run_capsule_ref).collect::<Vec<_>>();
-		let blockers = build_agent_blockers(&project_view, &blockers_dir, &run_refs);
+		let blockers =
+			agent_evidence::build_agent_blockers(&project_view, &blockers_dir, &run_refs);
 		let recovery_worktrees = project_view
 			.recovery_worktrees
 			.iter()
-			.map(|(role, worktree)| agent_recovery_worktree(role, worktree))
+			.map(|(role, worktree)| agent_evidence::agent_recovery_worktree(role, worktree))
 			.collect::<Vec<_>>();
 		let recovery_contracts = blockers.iter().filter_map(agent_recovery_contract).collect();
 		let connector_backoffs = project_view
@@ -90,7 +94,7 @@ pub(in crate::orchestrator) fn write_agent_evidence_snapshot(
 			events_path: &events_path,
 		};
 
-		write_agent_evidence_files(&write_context, &index, &run_capsules)?;
+		agent_evidence::write_agent_evidence_files(&write_context, &index, &run_capsules)?;
 
 		results.push(AgentEvidenceWriteResult {
 			project_id,
